- Automatic reflection file loading
- File:Exist: new routine
- Grouped texture export / ANSYS export
- Default view mode: Orthographic
- 0.1%BW spectrum implemented
- Auto-refresh B_factors every second
- Total flux/power: generated (instead of total absorbed)
- Better formula offsetting

+++ RegionInfo.cpp
@@ -264,7 +264,7 @@ void RegionInfo::Update() {
 	t4->SetText(tmp);
 	
 	if (selectedRegion->selectedPoint!=-1) {
-		Vector B=selectedRegion->B((double)selectedRegion->selectedPoint,Vector(0,0,0));
+		Vector B=selectedRegion->B(selectedRegion->selectedPoint,Vector(0,0,0));
 		sprintf(tmp,"B=%g T (%g , %g , %g)",B.Norme(),B.x,B.y,B.z);
 	} else sprintf(tmp,"");
 	t5->SetText(tmp);
